layout/forms/test/test_listcontrol_search.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=849438
-->
<head>
  <meta charset="utf-8">
  <title>Test for &lt;select&gt; list control search</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** This test will focus a select element and press a key that matches the
      first non-space character of an entry. **/

  SimpleTest.waitForExplicitFinish();
  SimpleTest.waitForFocus(function() {
    var select = document.getElementsByTagName('select')[0];
    select.focus();
    sendString('a');

    is(select.options[0].selected, false, "the first option isn't selected");
    is(select.options[1].selected, true, "the second option is selected");

    select.blur();

    SimpleTest.finish();
  });

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=849438">Mozilla Bug 849438</a>
<p id="display"></p>
<div id="content">
  <select>
    <option>Please select an entry</option>
    <option>&nbsp;a</option>
    <option>&nbsp;b</option>
  </select>
</div>
<pre id="test">
</pre>
</body>
</html>
